+++ core/domain/accounting/accountingcontroller.cpp
@@ -20,7 +20,7 @@
 **************************************************************************************************/
 #include "accountingcontroller.hpp"
 #include <memory>
-#include <general.hpp>  // general utility
+#include <generalutils.hpp>  // general utility
 #include <datetime/datetime.hpp>
 #include <logger/loghelper.hpp>
