+++ core/validator/productvalidator.cpp
@@ -20,7 +20,7 @@
 **************************************************************************************************/
 #include "productvalidator.hpp"
 #include <string>
-#include <general.hpp>  // pscore utility
+#include <generalutils.hpp>  // pscore utility
 
 namespace entity {
 namespace validator {
