+++ orchestra/application/screen/backoffice/inventoryscreen.cpp
@@ -25,7 +25,7 @@
 #include <memory>
 #include <string>
 #include <vector>
-#include <general.hpp>  // pscore utility
+#include <generalutils.hpp>  // pscore utility
 #include <generalhelper.hpp>  // view utility
 #include <informationscreen.hpp>
 #include <logger/loghelper.hpp>
