+++ utility/logger/filelog.cpp
@@ -23,7 +23,7 @@
 #include <iomanip>
 #include <sstream>
 #include <cfg/config.hpp>
-#include <general.hpp>
+#include <generalutils.hpp>
 
 namespace utility {
